// planning-idsdp-pps-frontend/src/app/main/modules/dpp-tapp-management/models/dppObjectiveCost.model.ts
export class DppObjectiveCostModel {
    id: number;
    uuid: string;
    status: string;
    projectTitleEn: string;
    projectTitleBn: string;
    ministryDivision: string;
    implementingAgency: string;
    objectivesTargets: string;
    concernedDivisionId: any;
    developmentPartnerId: number;
    dateCommencement: any;
    dateCompletion: any;
    projectConceptUuid: string;
    projectConceptMasterId: number;
    paripatraVersionId: number;
    dppMasterId: any;
    modeFinanceList:any;
    responsiblePreparation: string;
    designationContactPerson: string;
    fsUuid: string;
    fsAttachmentId: number;
    fsAttachmentName: string;
    sectorId: any;
    subSectorId: any;

   // dppModeOfFinancing: any;
   // currencyRateList: DppCurrencyRateModel[] = new Array<DppCurrencyRateModel>();
}